+++ Data/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## 1779667200
+## 1781481600
 
 ### vanilla
-- 1,376,278 entries as of May 25, 2026
+- 1,413,894 entries as of June 15, 2026
 
 ### tbc
-- 1,379,015 entries as of May 25, 2026
+- 1,416,853 entries as of June 15, 2026